refactor: ui styles
