+++ src/app/components/ui/common-styles.ts
@@ -1,2 +1,5 @@
-export const borderClasses =
+export const uiBorderClasses =
   "border-4 border-gray-800 p-2 focus:border-4 focus:border-blue-400 focus:outline-none";
+
+export const uiTextLondrinaWhite = "font-londrina-solid text-gray-200 text-xl";
+export const uiTextLondrinaBlack = "font-londrina-solid text-gray-200 text-xl";
